+++ src/components/Landing.vue
@@ -1,10 +1,20 @@
 <template>
   <v-container fluid>
     <v-layout row wrap>
-      <v-flex xs12 text-xs-center mt-5 mb-3>
-        <h1 class="display-2 font-weight-bold">{{ $t('landing.TITLE') }}</h1>
-      </v-flex>
-      <v-flex xs12 text-xs-center mb-2>{{ $t('landing.DESCRIPTION') }}</v-flex>
+      <Heading :title="$t('landing.TITLE')" />
+      <Description :description="$t('landing.DESCRIPTION')" />
     </v-layout>
   </v-container>
 </template>
+
+<script>
+import Heading from '@/components/common/Heading.vue'
+import Description from '@/components/common/Description.vue'
+
+export default {
+  components: {
+    Heading,
+    Description
+  }
+}
+</script>