+++ src/components/NotFound.vue
@@ -1,9 +1,17 @@
 <template>
   <v-container fluid>
     <v-layout row wrap>
-      <v-flex xs12 text-xs-center mt-5 mb-3>
-        <h1 class="display-2 font-weight-bold">{{ $t('errors.404') }}</h1>
-      </v-flex>
+      <Heading :title="$t('errors.404')" />
     </v-layout>
   </v-container>
 </template>
+
+<script>
+import Heading from '@/components/common/Heading.vue'
+
+export default {
+  components: {
+    Heading
+  }
+}
+</script>